Added simple user authentication.
Using Layers OpenID Connect server for authentication. The configuration
had to be a bit different than one for in-browser use (Grant Types: must
be "authorization code" and Response Types "code"). Redirect URL should
be <APP-URL>/auth/oidc/callback. Please make sure to add appropriate
clientID and clientSecret into the config.js file.
NB! The version of passport-openidconnect from github seems to be a bit
differnt than one in npm, maybe using the github repo directly is a good
idea.
NB! Using local variable for user storage. Has to be changed.
NB! Test do not reflect the behavioour of system with authenticated user
due to port being guessed for the tests (callbackURL) and not being sure
that the IDP has been configured for the tests. That would either need
to be manually mocked (pretend to have authenticated user) or some other
strategy has to be used instead.